FindBISON: Fix "multiple rules generate" error on VERBOSE output
When `BISON_TARGET` is called with both

    VERBOSE <extra-file> REPORT_FILE <file>

we add a custom command to copy `<file>` to `<extra-file>`.  The change
in commit v3.7.1~10^2 (FindBISON: Do not rebuild every time when not
VERBOSE, 2016-11-14) incorrectly listed the extra file as an output of
the main bison invocation custom command, but it is actually produced by
the separate copy command.  Fix the logic to always generate `<file>` as
an output of the bison command and `<extra-file>` as an output of the
copy command.

Fixes: #18433