FindBoost: Add explicit Boost_ARCHITECTURE option
Boost 1.66 and above built with `--layout=versioned` add an architecture
tag to the library file names.  We already try to compute this tag
automatically when `CMAKE_CXX_COMPILER_ARCHITECTURE_ID` is available,
but that is currently not computed everywhere.  Add an explicit
`Boost_ARCHITECTURE` option that a user can set to specify the
architecture tag.

Issue: #17701