VS: Add v140 flag table entries for `-Zc:inline[-]`
Note: Zc:inline was added by generator, but not the - version.

Original header commit v3.7.2~8^2

The documentation of this option [1] claims that the default is off, but
VS seems to use `-Zc:inline` by default if `RemoveUnreferencedCodeData`
does not appear in the `.vcxproj` file.  Add the flag table entry to
allow use of the flag to be configured.

[1] https://msdn.microsoft.com/en-us/library/dn642448.aspx