Autogen: Fix deadlock when uv_spawn() fails while trying to run moc
If by some chance the moc executable does not exist while running
AUTOMOC, instead of showing an error, the CMake Autogen invocation
hangs indefinitely.

This happens because UVProcessFinished() is not called if the process
does not launch correctly.

Make sure to call UVProcessFinished() even if the process launch fails,
and also report the error returned by libuv.