testing: Update terminology in ctest_update tests
Use the word "Update" instead of "Test" in ctest_update's RunCMake tests.

This change was prompted because I noticed that UpdateChangeId-check.cmake
was not actually getting called during the relevant test case.  This file
also had to be updated since its purpose is to detect an expected XML element
(not an attribute).